Refactored content sending method and updated tests (#484)
* "Refactored content sending method and updated tests"

The content sending logic has been refactored into a separate method called "SendContent" for better maintainability. This simplifies the HTTP response handling section in the "LocalHttpServer.cs" file, as the responsibility to position the stream, create a buffer, and write to the response output stream has been moved to this new helper method. The range of data to be sent is now calculated within this method itself.  Also, included Debug.Assert statement in "AdvancedHttpDownloader.cs" for debugging purpose.

* Move files out of /tmp/ on linux and into a folder closer to the tests

* Adds prefix to TemporaryFileManager in the CLI so that we don't use in-memory temp folders on Linux

* more prefixes

* change temp location

* Change temp location

* revert ubuntu version change

* Add retries to SendAsync

* fixes for the HTTP tests

* Standardize temp folder creation

* Add retry logic to handle SocketException in HttpDownloader

Added a try-catch block to handle SocketException in the process of reading response content in the HttpDownloader. This modification introduces a retry mechanism where, upon encountering a SocketException, the system will attempt to read the response content again. The number of retries is capped to avoid infinite retry loops. This enhancement increases the robustness of the HttpDownloader by providing a safeguard against unpredictable network errors.

* Can't resend requests, so copy it in the case of a retry