Move DiskStateRegistry to MnemonicDB (#1194)
Reworks the DiskStateRegistry to use MnemonicDB.

This is a first step to integrating MnemonicDB into the app, and shook out a lot of the small bugs I had yet to encounter.

What this does:

Creates 4 attributes for the four paramters of the DiskStateAchive
4 of these attributes are marked as 'noHistory:true' as we only ever want the most recent value, so no reason to track changes over time
All the attributes except for State are marked as IsIndexed=true meaning we can look up the state by any of these, or any combination of these attributes
Groups the attributes together into a model
Updates DiskStateRegistry to use this model
Adds an IConnection class into the app, and sets up everything to use MnemonicDB with the RocksDB backend
Disk state is stored in a FlatBuffer there's some nice properties this format offers us, read up on it if you're bored
MnemonicDB supports HashedBlobs now where the values can be super large (technically GB in size) and only the hash will be used for sorting, so we use that for storing the state
Not in this PR but to revisit later:

MnemonicDB needs to be updated to allow for entity paritions. We can get better performance by separating out entities into rough groups: loadouts, downloads, diskstate, settings, should likely be all in separate partitions. I need to add support to this to MnemonicDB. Once I do we can move DiskState to its own partition

We are hand crafting queries in this code currently (see the use of .FindIndexed) not a horrible issue, but we'll want a query engine eventually. I'm working on this

This PR resolves: #985 we no longer need to load the entire state to get the current applied loadout Id, later on we can query any of these values independantly of any others, simply by using .FindIndexed<>.