Fixed: Exception when same file ID is encountered multiple times.
This shouldn't usually happen, but it's technically possible if the user deletes multiple of the same mod in a single operation, leading to duplicates in the changeset.

For example, install 10 copies of SMAPI 4.2.0, and deleting them all at once.

We want to be resilient to those kinds of errors; so here we drop.