Fix "control already has a parent"-exception
Repro using the Library tree:

- Install a mod
- Uninstall the mod
- Install the mod again
- Get `InvalidOperationException`

The exception was thrown because we try to set a control as child of a
new border while it was still a child of the old border. This only
happens for fallback controls as they are only created once.

The PR properly detaches those controls before assigning them to
something else.